Move App Intents codes to separate files
